chore: 🔧 rename SideBar to sidebar
