+++ src/lib/components/navigation/Sidebar.svelte
